Fix Bugs
- PauseState problem
- missiles not firing due to wrong receivertypes
- missiles not firing in network code due to uint8_t being mixed
with int
- Outstanding issue of music not playing in the multiplayergamestate

+++ GD4_25_SFML/world.hpp
@@ -23,8 +23,8 @@ public:
 	sf::FloatRect GetViewBounds() const;
 	CommandQueue& GetCommandQueue();
 
-	Aircraft* AddAircraft(int identifier);
-	void RemoveAircraft(int identifier);
+	Aircraft* AddAircraft(uint8_t identifier);
+	void RemoveAircraft(uint8_t identifier);
 	void SetCurrentBattleFieldPosition(float line_y);
 	void SetWorldHeight(float height);
